% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.wrapper.r
\name{agg.wrapper}
\alias{agg.wrapper}
\title{Runs all the functions needed for aggregating the GGCMI-CMIP6 yields.}
\usage{
agg.wrapper(
  datafile,
  crop,
  region.map = "countries",
  custom.map = NULL,
  weights,
  type = "relative",
  irrigation = "total",
  is.ensemble = FALSE
)
}
\arguments{
\item{datafile}{Character string with the name of a RData file
provided by Jonas. May include a path e.g
"../data/dssat-pythia_gfdl-esm4_ssp126_default_production_and_yield_grid.RData"}

\item{crop}{One of "maize", "soybeans", "wheat", or "rice".}

\item{region.map}{A regional mapping. Current options are
"countries", "regionsGTAPV10.1" and "GTAPV11AEZREGS" or
"custom". If "custom" is chosen, a custom map should be
provided. Default is "countries". (See
\link[GGCMIAGG]{grid.agg}).}

\item{custom.map}{A regional mapping from gridcells to user-defined
region with columns labeled lon, lat, and id. (See
\link[GGCMIAGG]{grid.agg})}

\item{weights}{One of "none", "area", "production".}

\item{type}{Either "relative" (default) or "physical". Relative are
percentage changes relative to the reference period
1983-2013. Physical are in MT/ha.}

\item{irrigation}{Either "total" (default), "rainfed", or
"irrigated". It only matters when type="physical".}

\item{is.ensemble}{set to TRUE if aggregating ensemble means. By
default is FALSE.}
}
\value{
The output of \link[GGCMIAGG]{grid.agg}, that is, a dataframe with three
columns: countries (or other regional identifier), year, and
the grid-cell weighted average of relative yields.
}
\description{
The function sequentially runs the three functions needed to get
the aggregate yields: \link[GGCMIAGG]{read.weights},
\link[GGCMIAGG]{read.AgMIP.nc},  and
\link[GGCMIAGG]{grid.agg}.
}
