% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.agg.r
\name{grid.agg}
\alias{grid.agg}
\title{Aggregates GGCMI yield data to regions}
\usage{
grid.agg(
  data2agg = NULL,
  region.map = NULL,
  custom.map = NULL,
  weight.map = NULL
)
}
\arguments{
\item{data2agg}{A dataframe produced by \code{read.GGCMI.RData()}}

\item{region.map}{A regional mapping. Current options are
"countries", "regionsGTAPV10.1" and "countriesAEZ18" or
"custom". If "custom" is chosen, a custom map should be
provided. Default is "countries".}

\item{custom.map}{A regional mapping from gridcells to user-defined
region with columns labeled lon, lat, and id.}

\item{weight.map}{Either NULL (when 'weights' = "none" in
'agg.wrapper()') or the output of 'read.weights()'}
}
\value{
A dataframe with three columns: countries (or other
regional identifier), year, and the grid-cell weighted average
of relative yields.
}
\description{
This function should work with either CMIP5 or CMIP6 yields.
}
\details{
Aggregates from the gridcell to the regions in region.map. The
function \link[GGCMIAGG]{read.AgMIP.nc} produces the input to
\code{data2agg}. \link[GGCMIAGG]{read.weights} produces the data for teh
argument \code{weight.map}.
}
