% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AgMIP.nc.r
\name{read.AgMIP.nc}
\alias{read.AgMIP.nc}
\title{Reads GGCMI yields from netcdf files}
\usage{
read.AgMIP.nc(file, start_year, end_year, yield_crop, var_lon, var_lat)
}
\arguments{
\item{file}{A NetCDF file with GGCMI yields}

\item{start_year}{First year to be extracted.}

\item{end_year}{Last year to be extracted.}

\item{yield_crop}{The name of the variable containing the yields in
the netcdf file. In the CMIP 5 yields it was a crop name. In
the CMIP 6 yiels it's 'yield cahange'}

\item{var_lon}{Longitude variable in the NetCDF file.}

\item{var_lat}{Latitude variable in the NetCDF file.}
}
\value{
A dataframe with four columns: lon, lat, year, and
projected.yield. NAs are eliminated.
}
\description{
The function read.AgMIP.nc takes as an argument a file
name(character string) that identifies a unique ntcdf file, start
year, end year and variable id for yield, longitude, and latitude.
Look into the header of the ncdf4 file for these variables.  The
function converts the selected ncdf file into a table with four
columns: lon, lat, year, and projected.yield. NAs are eliminated.
This function is not used for the version of the tool using the
second generation of the GGCMI runs, but it is kept for legacy.
This function should work with either CMIP5 or CMIP6 yields.
}
