% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.GGCMI.RData.r
\name{read.GGCMI.RData}
\alias{read.GGCMI.RData}
\title{Reads RData files with GGCMI-CMIP6 yields in RData format}
\usage{
read.GGCMI.RData(datafile = NULL, crop = NULL)
}
\arguments{
\item{datafile}{Character string with the name of a RData file
provided by Jonas, inclusive of path e.g
"../data/dssat-pythia_gfdl-esm4_ssp126_default_production_and_yield_grid.RData"}

\item{crop}{One of "maize", "winter_wheat", "spring_wheat",
"soybeans", "wheat", or "rice".}
}
\value{
A four-column dataframe
("lon","lat","time","value"). "time" are all the years in the
dataset (e.g., for future runs 2016:2099), and value are the
relative yields for \code{crop}. The data is ready to be aggregated
geographically by \code{grid.agg()}, which merges the lon/lat
variables to a lower resolution with the weights produced by
'read.weights()`.
}
\description{
Reads yield data (relative changes in percent compared with the
reference period 1983-2013) from the CMIP6 generation of GGCMI runs
provided by Jonas Jagermeyr on June 01, 2022. See \href{https://www.nature.com/articles/s43016-021-00400-y}{Jagermeyr et al.} for
description.
}
