% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.wrapper.r
\name{agg.wrapper}
\alias{agg.wrapper}
\title{Runs all the functions needed for aggregating the GGCMI yields.}
\usage{
agg.wrapper(datafile, crop, regions = "countries", weights)
}
\arguments{
\item{datafile}{Character string with the name of a RData file
provided by Jonas, inclusive of path e.g
"../data/dssat-pythia_gfdl-esm4_ssp126_default_production_and_yield_grid.RData"}

\item{crop}{One of "maize", "winter_wheat", "spring_wheat",
"soybeans", or "rice".}

\item{regions}{Current options are are "countries" and
"countriesAEZ18". Default is "countries".}

\item{weights}{Either NULL (when 'weights' = "none" in
'agg.wrapper()') or the output of 'read.weights()'}
}
\value{
The output of \link[GGCMIAGG]{grid.agg}, that is, a dataframe with three
columns: countries (or other regional identifier), year, and
the grid-cell weighted average of relative yields.
}
\description{
The function sequentially runs the three functions needed to get
the aggregate yields.
}
