% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AgMIP.nc.r
\name{read.AgMIP.nc}
\alias{read.AgMIP.nc}
\title{Reads GGCMI yields from netcdf datafiles}
\usage{
read.AgMIP.nc(
  datafile,
  start_year = 1983,
  end_year = 2099,
  nc4varid = "yield change",
  targetcrop,
  cropsinnc = c("maize", "wheat", "soybeans", "rice"),
  var_lon = "lon",
  var_lat = "lat"
)
}
\arguments{
\item{datafile}{A NetCDF datafile with GGCMI yields for four crops
(see cropsinnc) below, for 0.5 degree gridcells and years
1983:2099. These are percentage changes in yields relative to
1983-2013. These were provided by AgMIP GRIDded crop modeling
initiative (Ag-GRID) coleads Jonas Jaegermeyr and Christoph
Mueller.}

\item{start_year}{First year to be extracted. Defaults to 1983.}

\item{end_year}{Last year to be extracted. Defaults to 2099.}

\item{nc4varid}{The name of the variable containing the yields in
the netcdf datafile. Defaults to "yield change"}

\item{targetcrop}{the crop (one of cropsinnc below) for which the
change in relative yields are requested.}

\item{cropsinnc}{Available crops. Defaults to "maize", "wheat",
"soybeans" and "rice".}

\item{var_lon}{Longitude variable in the NetCDF datafile. Defaults
to "lon".}

\item{var_lat}{Latitude variable in the NetCDF datafile. Defaults
to "lat".}
}
\value{
A four-column dataframe
("lon","lat","time","value"). "time" are all the years in the
dataset (e.g., for future runs 2016:2099), and value are the
relative yields for \code{crop}. The data is ready to be aggregated
geographically by \code{grid.agg()}, which merges the lon/lat
variables to a lower resolution with the weights produced by
'read.weights()`.
}
\description{
Reads yield data (relative changes in percent compared with the
reference period 1983-2013) from the CMIP6 generation of GGCMI runs
provided by Jonas Jaegermeyr on December 9th, 2022. See \href{https://www.nature.com/articles/s43016-021-00400-y}{Jaegermeyr et al.} for
description.
}
